refactor: finally move the pausable_failpoint
