+++ pageserver/src/tenant/remote_timeline_client/upload.rs
@@ -34,7 +34,7 @@ pub(super) async fn upload_index_part<'a>(
     fail_point!("before-upload-index", |_| {
         bail!("failpoint before-upload-index")
     });
-    pausable_failpoint!("before-upload-index-pausable");
+    crate::failpoint_support::pausable_failpoint!("before-upload-index-pausable");
 
     let index_part_bytes = index_part
         .to_s3_bytes()
@@ -68,7 +68,7 @@ pub(super) async fn upload_timeline_layer<'a>(
         bail!("failpoint before-upload-layer")
     });
 
-    pausable_failpoint!("before-upload-layer-pausable");
+    crate::failpoint_support::pausable_failpoint!("before-upload-layer-pausable");
 
     let storage_path = remote_path(conf, source_path, generation)?;
     let source_file_res = fs::File::open(&source_path).await;
